+++ Cookiee/Cookiee/Presentation/Calendar/View/CalendarControlFunctions.swift
@@ -1,0 +1,102 @@
+//
+//  CalendarControllFunctions.swift
+//  Cookiee
+//
+//  Created by minseo Kyung on 10/15/24.
+//
+
+import SwiftUI
+
+extension HomeCalendarView {
+    var today: Date {
+        let now = Date()
+        let components = Calendar.current.dateComponents([.year, .month, .day], from: now)
+        return Calendar.current.date(from: components)!
+    }
+  
+    static let calendarHeaderDateFormatter: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "YYYY년 MM월"
+        return formatter
+    }()
+  
+    static let weekdaySymbols: [String] = Calendar.current.shortWeekdaySymbols
+    
+    func getDate(for index: Int) -> Date {
+        let calendar = Calendar.current
+        guard let firstDayOfMonth = calendar.date(
+            from: DateComponents(
+            year: calendar.component(.year, from: month),
+            month: calendar.component(.month, from: month),
+            day: 1
+            )
+        ) else {
+            return Date()
+        }
+    
+        var dateComponents = DateComponents()
+        dateComponents.day = index
+        
+        let timeZone = TimeZone.current
+        let offset = Double(timeZone.secondsFromGMT(for: firstDayOfMonth))
+        dateComponents.second = Int(offset)
+        
+        let date = calendar.date(byAdding: dateComponents, to: firstDayOfMonth) ?? Date()
+        return date
+    }
+      
+    func numberOfDays(in date: Date) -> Int {
+        return Calendar.current.range(of: .day, in: .month, for: date)?.count ?? 0
+    }
+  
+    func firstWeekdayOfMonth(in date: Date) -> Int {
+        let components = Calendar.current.dateComponents([.year, .month], from: date)
+        let firstDayOfMonth = Calendar.current.date(from: components)!
+
+        return Calendar.current.component(.weekday, from: firstDayOfMonth)
+    }
+
+    func previousMonth() -> Date {
+        let components = Calendar.current.dateComponents([.year, .month], from: month)
+        let firstDayOfMonth = Calendar.current.date(from: components)!
+        let previousMonth = Calendar.current.date(byAdding: .month, value: -1, to: firstDayOfMonth)!
+
+        return previousMonth
+    }
+
+    func changeMonth(by value: Int) {
+        self.month = adjustedMonth(by: value)
+    }
+
+    func adjustedMonth(by value: Int) -> Date {
+        if let newMonth = Calendar.current.date(byAdding: .month, value: value, to: month) {
+            return newMonth
+        }
+        return month
+    }
+    
+    func filterThumbnailUrlByDate(date: Date) -> String? {
+        let calendar = Calendar.current
+        let year = calendar.component(.year, from: date)
+        let month = calendar.component(.month, from: date)
+        let day = calendar.component(.day, from: date)
+
+        return homeCalendarViewModel.thumbnailList.first(where: {
+            $0.eventYear == year &&
+            $0.eventMonth == month &&
+            $0.eventDate == day
+        })?.thumbnailUrl
+    }
+}
+
+extension Date {
+    static let calendarDayDateFormatter: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateFormat = "MMMM yyyy dd"
+        return formatter
+    }()
+  
+    var formattedCalendarDayDate: String {
+        return Date.calendarDayDateFormatter.string(from: self)
+    }
+}
