+++ Cookiee/Cookiee/Presentation/Calendar/View/HomeCalendarView.swift
@@ -8,8 +8,8 @@
 import SwiftUI
 
 struct HomeCalendarView: View {
-    @State private var month: Date = Date()
-    @State private var clickedCurrentMonthDates: Date?
+    @State var month: Date = Date()
+    @State var clickedCurrentMonthDates: Date?
     
     @ObservedObject var homeCalendarViewModel = CalendarThumnailViewModel()
     
@@ -24,7 +24,6 @@ struct HomeCalendarView: View {
     var body: some View {
         GeometryReader { geometry in
             VStack {
-                // 쿠키 타이틀 헤더
                 ZStack {
                     Image("cookiee_typo_small")
                     HStack {
@@ -38,7 +37,6 @@ struct HomeCalendarView: View {
                 }
                 .frame(width: geometry.size.width, height: 45)
                 
-                // 캘린더
                 VStack {
                     headerView
                     calendarGridView
@@ -52,7 +50,6 @@ struct HomeCalendarView: View {
         }
     }
     
-    // 헤더
     private var headerView: some View {
         VStack(alignment: .center) {
             yearMonthView
@@ -72,7 +69,6 @@ struct HomeCalendarView: View {
         .frame(height: 75)
     }
 
-    // 연월 표기
     private var yearMonthView: some View {
         HStack {
             Button(
@@ -107,152 +103,46 @@ struct HomeCalendarView: View {
     }
 
       
-      // 캘린더 그리드 뷰
-      private var calendarGridView: some View {
+    private var calendarGridView: some View {
         let daysInMonth: Int = numberOfDays(in: month)
         let firstWeekday: Int = firstWeekdayOfMonth(in: month) - 1
         let lastDayOfMonthBefore = numberOfDays(in: previousMonth())
         let numberOfRows = Int(ceil(Double(daysInMonth + firstWeekday) / 7.0))
         let visibleDaysOfNextMonth = numberOfRows * 7 - (daysInMonth + firstWeekday)
         
-          return LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 0, alignment: .top), count: 7), spacing: 2) {
-          ForEach(-firstWeekday ..< daysInMonth + visibleDaysOfNextMonth, id: \.self) { index in
-            Group {
-              if index > -1 && index < daysInMonth {
-                let date = getDate(for: index)
-                let day = Calendar.current.component(.day, from: date)
-                let clicked = clickedCurrentMonthDates == date
-                let isToday = date.formattedCalendarDayDate == today.formattedCalendarDayDate
-                  
-                  let thumbnailUrl = filterThumbnailUrlByDate(date: date)
-                                
-                  NavigationLink(
-                      destination: DateView(date: date),
-                      label: {
-                          CellView(day: day, clicked: clicked, isToday: isToday, thumbnailUrl: thumbnailUrl)
-                      }
-                  )
-              } else if let prevMonthDate = Calendar.current.date(
-                byAdding: .day,
-                value: index + lastDayOfMonthBefore,
-                to: previousMonth()
-              ) {
-                let day = Calendar.current.component(.day, from: prevMonthDate)
-                  CellView(day: day, isCurrentMonthDay: false, thumbnailUrl: nil)
-              }
+        return LazyVGrid(columns: Array(repeating: GridItem(.flexible(), spacing: 0, alignment: .top), count: 7), spacing: 2) {
+            ForEach(-firstWeekday ..< daysInMonth + visibleDaysOfNextMonth, id: \.self) { index in
+                Group {
+                    if index > -1 && index < daysInMonth {
+                        let date = getDate(for: index)
+                        let day = Calendar.current.component(.day, from: date)
+                        let clicked = clickedCurrentMonthDates == date
+                        let isToday = date.formattedCalendarDayDate == today.formattedCalendarDayDate
+                        
+                        let thumbnailUrl = filterThumbnailUrlByDate(date: date)
+                        
+                        NavigationLink(
+                            destination: DateView(date: date),
+                            label: {
+                                CellView(day: day, clicked: clicked, isToday: isToday, thumbnailUrl: thumbnailUrl)
+                            }
+                        )
+                    } else if let prevMonthDate = Calendar.current.date(
+                        byAdding: .day,
+                        value: index + lastDayOfMonthBefore,
+                        to: previousMonth()
+                    ) {
+                        let day = Calendar.current.component(.day, from: prevMonthDate)
+                        CellView(day: day, isCurrentMonthDay: false, thumbnailUrl: nil)
+                    }
+                }
+                .onTapGesture {
+                    if 0 <= index && index < daysInMonth {
+                        let date = getDate(for: index)
+                        clickedCurrentMonthDates = date
+                    }
+                }
             }
-            .onTapGesture {
-              if 0 <= index && index < daysInMonth {
-                let date = getDate(for: index)
-                clickedCurrentMonthDates = date
-              }
-            }
-          }
         }
     }
-}
-
-// MARK: - 날짜에 맞는 썸네일 URL 반환
-private extension HomeCalendarView {
-    func filterThumbnailUrlByDate(date: Date) -> String? {
-        let calendar = Calendar.current
-        let year = calendar.component(.year, from: date)
-        let month = calendar.component(.month, from: date)
-        let day = calendar.component(.day, from: date)
-
-        return homeCalendarViewModel.thumbnailList.first(where: {
-            $0.eventYear == year &&
-            $0.eventMonth == month &&
-            $0.eventDate == day
-        })?.thumbnailUrl
-    }
-}
-
-
-private extension HomeCalendarView {
-  var today: Date {
-    let now = Date()
-    let components = Calendar.current.dateComponents([.year, .month, .day], from: now)
-    return Calendar.current.date(from: components)!
-  }
-  
-  static let calendarHeaderDateFormatter: DateFormatter = {
-    let formatter = DateFormatter()
-    formatter.dateFormat = "YYYY년 MM월"
-    return formatter
-  }()
-  
-  static let weekdaySymbols: [String] = Calendar.current.shortWeekdaySymbols
-}
-
-private extension HomeCalendarView {
-  /// 특정 해당 날짜
-  func getDate(for index: Int) -> Date {
-    let calendar = Calendar.current
-    guard let firstDayOfMonth = calendar.date(
-      from: DateComponents(
-        year: calendar.component(.year, from: month),
-        month: calendar.component(.month, from: month),
-        day: 1
-      )
-    ) else {
-      return Date()
-    }
-    
-    var dateComponents = DateComponents()
-    dateComponents.day = index
-    
-    let timeZone = TimeZone.current
-    let offset = Double(timeZone.secondsFromGMT(for: firstDayOfMonth))
-    dateComponents.second = Int(offset)
-    
-    let date = calendar.date(byAdding: dateComponents, to: firstDayOfMonth) ?? Date()
-    return date
-  }
-  
-  /// 해당 월에 존재하는 일자 수
-  func numberOfDays(in date: Date) -> Int {
-    return Calendar.current.range(of: .day, in: .month, for: date)?.count ?? 0
-  }
-  
-  /// 해당 월의 첫 날짜가 갖는 해당 주의 몇번째 요일
-  func firstWeekdayOfMonth(in date: Date) -> Int {
-    let components = Calendar.current.dateComponents([.year, .month], from: date)
-    let firstDayOfMonth = Calendar.current.date(from: components)!
-    
-    return Calendar.current.component(.weekday, from: firstDayOfMonth)
-  }
-  
-  /// 이전 월 마지막 일자
-  func previousMonth() -> Date {
-    let components = Calendar.current.dateComponents([.year, .month], from: month)
-    let firstDayOfMonth = Calendar.current.date(from: components)!
-    let previousMonth = Calendar.current.date(byAdding: .month, value: -1, to: firstDayOfMonth)!
-    
-    return previousMonth
-  }
-  
-  /// 월 변경
-  func changeMonth(by value: Int) {
-    self.month = adjustedMonth(by: value)
-  }
-  
-  /// 변경하려는 월 반환
-  func adjustedMonth(by value: Int) -> Date {
-    if let newMonth = Calendar.current.date(byAdding: .month, value: value, to: month) {
-      return newMonth
-    }
-    return month
-  }
-}
-extension Date {
-  static let calendarDayDateFormatter: DateFormatter = {
-    let formatter = DateFormatter()
-    formatter.dateFormat = "MMMM yyyy dd"
-    return formatter
-  }()
-  
-  var formattedCalendarDayDate: String {
-    return Date.calendarDayDateFormatter.string(from: self)
-  }
 }
